wayland: Add basic colorimetrie support
Using the Wayland color-management and color-representation protocol
drafts.

On the Gst caps side it is currently limited to a subset of predefined
GstVideoColorimetrys: sRGB, bt601, bt709, bt2020-10, bt2100-pq and
bt2100-hlq.

On the Wayland side it uses named transfer functions, named primaries,
matrices, ranges and chroma locations - hopefully with a correct
matching.

The implementation queries supported values from the compositors and tries
to convert them into known GstVideoColorimetrys. It currently *does not*
pass these upstream to decoders etc. as GstCaps for negotiation.

On setting caps it translates the GstVideoColorimetry from the GstVideoInfo
back to into a Wayland parametric image description and color representation
for the video surface if possible. If a colorimetry is not fully
support, we bail out and if wayland objects already exist they get reset or
deleted.