timecodestamper: Add scale property
If set to true, it will scale an incoming timecode of a potentially
different framerate than our current one (e.g. if we change frame rates
upstream of the timecodestamper).

For auto_resync, it will make sure that there's not more than 1 frame of
difference between upstream and framerate-scaled timecode.